update testnet endpoints
