+++ packages/hashes.csv
@@ -21,7 +21,7 @@ fetchai/agents/weather_station,QmYQCKmgaLDRnUtwJZEN1AYgEpFm2RYRxzB9LyT5iUJMx2
 fetchai/connections/gym,QmZUXKDqpyYPJo66CMHGRHF8LgqPtn4SV8CN3uTCaHq5ww
 fetchai/connections/http_client,QmfSuWhrseBCVcmzwKmqFDC6qX3ryFzfx2uuUTb8HPGuEd
 fetchai/connections/http_server,Qmc4cy6QcGDz6eeE5RCMWy5DkWnQqycTgLKQro9wM4qxFZ
-fetchai/connections/ledger,QmeYfof91qgXbv4tMJq2VoEVscXS1Mcx9Q35wJyVCXzHZk
+fetchai/connections/ledger,QmXHSjaQDb2PxEwhFy4ff3z1nEqXXR7VFU1c9jYxDDqFjA
 fetchai/connections/local,QmdgTUMdmjyuy6hXzS8CyZXWTPdiiEbjMudTFmW1rHRQ18
 fetchai/connections/oef,QmeXkrmmn847Nh6mzitcS4J6ixtGY2A757HWkhPdHpfyvi
 fetchai/connections/p2p_libp2p,QmbMd41K6HFWddjQiqcKXYNk26bfFcNvDRJuiDY3X9Tdj9
